feat: refactored useDocumentPiP and  add JSDoc comments to Document PiP composables [WTEL-9414](https://webitel.atlassian.net/browse/WTEL-9414)
