+++ src/modules/CallSession/modules/VideoCall/composables/useDocumentPiP/index.ts
@@ -1,0 +1,1 @@
+export { useDocumentPiP } from './useDocumentPiP';
